+++ wiki/Front-End/example/limitations/comments.jsx
@@ -1,0 +1,7 @@
+import { render }  from 'preact'
+
+const example = 'hello'
+render(<div>
+  {/* hello world */}
+  {/* <span>Comment {example}</span> */}
+</div>)