+++ wiki/Front-End/example/limitations/gt.jsx
@@ -1,0 +1,4 @@
+// won't work
+render(<div title="hello > world">
+  {(length > 10) && <span>Next Page</span>}
+</div>)