+++ wiki/Front-End/frontend/Example.jsx
@@ -1,0 +1,10 @@
+import { Component } from 'preact'
+import { $Example } from './style.css'
+
+export default class Example extends Component {
+  render() {
+    return (<div className={$Example}>
+      Idio Web Server.
+    </div>)
+  }
+}