+++ wiki/Front-End/frontend/index.jsx
@@ -1,0 +1,4 @@
+import { render } from 'preact'
+import Example from './Example'
+
+render(Example, document.body)